refactor(devtools): move .jsonl recording codec from standalone to bridge
