+++ redux-kotlin-devtools-bridge/api/jvm/redux-kotlin-devtools-bridge.api
@@ -238,3 +238,48 @@ public final class org/reduxkotlin/devtools/bridge/BridgeProtocolKt {
 	public static final fun toWire (Lorg/reduxkotlin/devtools/DevToolsEvent;)Lorg/reduxkotlin/devtools/bridge/BridgeMessage;
 }
 
+public final class org/reduxkotlin/devtools/bridge/RecordingHeader {
+	public static final field Companion Lorg/reduxkotlin/devtools/bridge/RecordingHeader$Companion;
+	public fun <init> (Ljava/lang/String;ILjava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)V
+	public synthetic fun <init> (Ljava/lang/String;ILjava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun component1 ()Ljava/lang/String;
+	public final fun component2 ()I
+	public final fun component3 ()Ljava/lang/String;
+	public final fun component4 ()Ljava/lang/String;
+	public final fun component5 ()Ljava/lang/String;
+	public final fun component6 ()Ljava/lang/String;
+	public final fun component7 ()Ljava/lang/String;
+	public final fun copy (Ljava/lang/String;ILjava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)Lorg/reduxkotlin/devtools/bridge/RecordingHeader;
+	public static synthetic fun copy$default (Lorg/reduxkotlin/devtools/bridge/RecordingHeader;Ljava/lang/String;ILjava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;ILjava/lang/Object;)Lorg/reduxkotlin/devtools/bridge/RecordingHeader;
+	public fun equals (Ljava/lang/Object;)Z
+	public final fun getClientId ()Ljava/lang/String;
+	public final fun getClientLabel ()Ljava/lang/String;
+	public final fun getKind ()Ljava/lang/String;
+	public final fun getProtocolVersion ()I
+	public final fun getSerializerTier ()Ljava/lang/String;
+	public final fun getStoreInstanceId ()Ljava/lang/String;
+	public final fun getStoreName ()Ljava/lang/String;
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
+}
+
+public final synthetic class org/reduxkotlin/devtools/bridge/RecordingHeader$$serializer : kotlinx/serialization/internal/GeneratedSerializer {
+	public static final field INSTANCE Lorg/reduxkotlin/devtools/bridge/RecordingHeader$$serializer;
+	public final fun childSerializers ()[Lkotlinx/serialization/KSerializer;
+	public synthetic fun deserialize (Lkotlinx/serialization/encoding/Decoder;)Ljava/lang/Object;
+	public final fun deserialize (Lkotlinx/serialization/encoding/Decoder;)Lorg/reduxkotlin/devtools/bridge/RecordingHeader;
+	public final fun getDescriptor ()Lkotlinx/serialization/descriptors/SerialDescriptor;
+	public synthetic fun serialize (Lkotlinx/serialization/encoding/Encoder;Ljava/lang/Object;)V
+	public final fun serialize (Lkotlinx/serialization/encoding/Encoder;Lorg/reduxkotlin/devtools/bridge/RecordingHeader;)V
+	public fun typeParametersSerializers ()[Lkotlinx/serialization/KSerializer;
+}
+
+public final class org/reduxkotlin/devtools/bridge/RecordingHeader$Companion {
+	public final fun serializer ()Lkotlinx/serialization/KSerializer;
+}
+
+public final class org/reduxkotlin/devtools/bridge/RecordingKt {
+	public static final fun decodeRecording (Ljava/lang/String;)Lkotlin/Pair;
+	public static final fun encodeRecording (Lorg/reduxkotlin/devtools/bridge/RecordingHeader;Ljava/util/List;)Ljava/lang/String;
+}
+
